chore(DTFS2-6994): rename and move unit tests labelled as api tests
