+++ portal-api/src/v1/validation/validateIds.test.js
@@ -1,5 +1,12 @@
-const { isValidMongoId, isValidRegex, isValidCurrencyCode, isValidCompaniesHouseNumber, isValidMonth, isValidYear } = require('../../../src/v1/validation/validateIds');
-const { COMPANIES_HOUSE_NUMBER } = require('../../fixtures/companies-house-number');
+const {
+  isValidMongoId,
+  isValidRegex,
+  isValidCurrencyCode,
+  isValidCompaniesHouseNumber,
+  isValidMonth,
+  isValidYear 
+} = require('./validateIds');
+const { COMPANIES_HOUSE_NUMBER } = require('../../../api-tests/fixtures/companies-house-number');
 
 const { VALID, VALID_LAST_LETTER, VALID_LETTERS, VALID_LETTERS_NI, INVALID_SHORT, INVALID_SPECIAL_CHARACTER, INVALID_SPACE } = COMPANIES_HOUSE_NUMBER;
